+++ eth2near/eth2near-block-relay-rs/config_for_tests.toml
@@ -1,13 +1,11 @@
 beacon_endpoint = "https://lodestar-goerli.chainsafe.io"
-eth1_endpoint = "https://goerli.infura.io/v3/64aa9edc46d941c3b30882ad2eae4304"
+eth1_endpoint = "https://goerli.infura.io/v3/API_KEY"
 path_to_current_sync_committee = "./data/goerli_testdata/next_sync_committee_goerli_period_473.json"
 path_to_next_sync_committee = "./data/goerli_testdata/next_sync_committee_goerli_period_474.json"
 path_to_execution_blocks_headers = "./data/goerli_testdata/execution_block_headers_goerli_slots_3885697_3886176.json"
 path_to_light_client_updates = "./data/goerli_testdata/light_client_updates_goerli_slots_3885697_3886176.json"
 path_to_attested_state = "./data/goerli_testdata/beacon_state_goerli_slot_3885731.json"
 path_to_finality_state = "./data/goerli_testdata/beacon_state_goerli_slot_3885664.json"
-path_to_attested_state_for_period = "./data/goerli_testdata/beacon_state_goerli_slot_3887699.json"
-path_to_finality_state_for_period = "./data/goerli_testdata/beacon_state_goerli_slot_3887616.json"
 network_name = "Goerli"
 first_slot = 3885632
 slot_without_block = 3885649
@@ -17,7 +15,3 @@ right_empty_slot = 3885723
 slot_without_block_2 = 3885743
 wasm_filepath = "../../contracts/near/res/eth2_client.wasm"
 finalized_slot_before_new_period = 3891168
-eth1_number = 0x73cb47
-path_to_block = "./data/goerli_testdata/beacon_block_goerli_slot_3885632.json"
-path_to_header = "./data/goerli_testdata/beacon_header_goerli_slot_3885632.json"
-path_to_light_client_update = "./data/goerli_testdata/beacon_light_client_update_goerli_period_474.json"
