--- eth2near/eth2near-block-relay-rs/data/goerli_testdata/beacon_block_goerli_slot_3885632.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d2a0aa876cbdc25468ca840d5b39d8357794df30d53210928e33e72916fe1b9a
-size 88130
