--- eth2near/eth2near-block-relay-rs/data/goerli_testdata/beacon_header_goerli_slot_3885632.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:57b25936d9f8f3e2481fc9b4a8b92155568d5ef06935d933e2eeeb3c8f925152
-size 624
