--- eth2near/eth2near-block-relay-rs/data/goerli_testdata/beacon_light_client_update_goerli_period_474.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e3aacd1473e5c19e1204aed56b4eda643effeaa6a602299af3f440c0ebfef2ee
-size 53734
