+++ eth2near/eth2near-block-relay-rs/src/config_for_tests.rs
@@ -22,12 +22,6 @@ pub struct ConfigForTests {
     pub right_empty_slot: u64,
     pub finalized_slot_before_new_period: u64,
     pub slot_without_block_2: u64,
-    pub path_to_attested_state_for_period: String,
-    pub path_to_finality_state_for_period: String,
-    pub eth1_number: u64,
-    pub path_to_block: String,
-    pub path_to_header: String,
-    pub path_to_light_client_update: String,
 }
 
 impl ConfigForTests {
