+++ eth2near/eth2near-block-relay-rs/src/last_slot_searcher.rs
@@ -1,5 +1,5 @@
-use crate::beacon_rpc_client::BeaconRPCClient;
-use crate::relay_errors::{ExecutionPayloadError, NoBlockForSlotError};
+use eth_rpc_client::beacon_rpc_client::BeaconRPCClient;
+use eth_rpc_client::errors::{ExecutionPayloadError, NoBlockForSlotError};
 use contract_wrapper::eth_client_contract_trait::EthClientContractTrait;
 use eth_types::H256;
 use log::{debug, trace};
@@ -398,9 +398,9 @@ impl LastSlotSearcher {
 
 #[cfg(test)]
 mod tests {
-    use crate::beacon_rpc_client::BeaconRPCClient;
+    use eth_rpc_client::beacon_rpc_client::BeaconRPCClient;
     use crate::config_for_tests::ConfigForTests;
-    use crate::eth1_rpc_client::Eth1RPCClient;
+    use eth_rpc_client::eth1_rpc_client::Eth1RPCClient;
     use crate::last_slot_searcher::LastSlotSearcher;
     use crate::test_utils::get_client_contract;
     use contract_wrapper::eth_client_contract_trait::EthClientContractTrait;
