+++ eth2near/eth2near-block-relay-rs/src/lib.rs
@@ -1,20 +1,14 @@
 extern crate core;
 
-pub mod beacon_block_body_merkle_tree;
-pub mod beacon_rpc_client;
 pub mod config;
-mod contract_type;
-pub mod eth1_rpc_client;
+pub mod contract_type;
 pub mod eth2near_relay;
-pub mod execution_block_proof;
-pub mod hand_made_finality_light_client_update;
 pub mod last_slot_searcher;
 pub mod logger;
-pub mod relay_errors;
+pub mod prometheus_metrics;
 
 #[cfg(test)]
 pub mod config_for_tests;
 
-pub mod prometheus_metrics;
 #[cfg(test)]
 pub mod test_utils;
