+++ eth2near/eth_rpc_client/Cargo.toml
@@ -1,7 +1,9 @@
 [package]
-name = "eth2_to_near_relay"
+name = "eth_rpc_client"
 version = "0.1.0"
 edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 log = { version = "0.4", features = ["std", "serde"] }
@@ -15,12 +17,10 @@ ethereum-types = "0.9.2"
 reqwest = { version = "0.11", features = ["blocking"] }
 eth-types = { path = "../../contracts/near/eth-types/", features = ["eip1559"]}
 contract_wrapper = { path = "../contract_wrapper" }
-eth_rpc_client = { path = "../eth_rpc_client" }
 clap = { version = "3.1.6", features = ["derive"] }
 tokio = { version = "1.1", features = ["macros", "rt", "time"] }
 env_logger = "0.9.0"
 borsh = "0.9.3"
-near-sdk = "4.0.0"
 futures = { version = "0.3.21", default-features = false }
 async-std = "1.12.0"
 hex = "*"
@@ -29,20 +29,11 @@ finality-update-verify = { path = "../finality-update-verify" }
 atomic_refcell = "0.1.8"
 bitvec = "*"
 
-near-jsonrpc-client = "0.4.0-beta.0"
-near-crypto = "0.14.0"
-near-primitives = "0.14.0"
-near-chain-configs = "0.14.0"
-near-jsonrpc-primitives = "0.14.0"
-
 prometheus = { version = "0.9", features = ["process"] }
 lazy_static = "1.4"
 warp = "0.2"
 thread = "*"
 
-[dev-dependencies]
-workspaces = "0.5.0"
-eth2-contract-init = { path = "../eth2-contract-init" }
 
 [patch]
 [patch.crates-io]
