+++ eth2near/eth_rpc_client/data/beacon_block_header_struct_kiln_slot_741888.json
