+++ eth2near/eth_rpc_client/data/beacon_block_kiln_slot_741888.json
