+++ eth2near/eth_rpc_client/data/beacon_header_goerli_slot_3885632.json
@@ -1,0 +1,1 @@
+{"data":{"root":"0xa1600e330efcf934b23706365793966c4fed60d34dcdd897620810315b5abda8","canonical":true,"header":{"message":{"slot":"3885632","proposer_index":"121460","parent_root":"0x0cb13db52b63240d561b36495723229e0424fc72a25e74c02b61ae448a5f6347","state_root":"0xad73703cc94bfa83619e913bb7de2850ff08d613008a74a80df4e6977f23517c","body_root":"0x6b603cb237df9ebab0c7dc55d37ea99a2cfa381a49d220493df8f3cc3922905d"},"signature":"0x90aca2fff31ad0425a0f2a25bd92d948b5d34dfcddb15ea7fc3a7984708c70b28f531971d7b73fafd68d54ecb9e6275a0f7153c4e53c2a66b0a004338cf252e07d101e7807667bb456f38df5eff4a6b27cf1a1b51c70eb3af5297c484ad462be"}}}
