+++ eth2near/eth_rpc_client/data/beacon_state_goerli_slot_3887699.json
