+++ eth2near/eth_rpc_client/src/utils.rs
@@ -1,0 +1,20 @@
+pub fn trim_quotes(s: String) -> String {
+    let mut res_str = s;
+    if (res_str.starts_with('"') && res_str.ends_with('"'))
+        || (res_str.starts_with('\'') && res_str.ends_with('\''))
+    {
+        res_str.pop();
+        res_str.remove(0);
+    }
+
+    res_str
+}
+
+pub fn read_json_file_from_data_dir(file_name: &str) -> std::string::String {
+    let mut json_file_path = std::env::current_exe().unwrap();
+    json_file_path.pop();
+    json_file_path.push("../../../data");
+    json_file_path.push(file_name);
+
+    std::fs::read_to_string(json_file_path).expect("Unable to read file")
+}
